Deleted useless classes and performed a better package organization.
Added a second container (fragmentContainer child of main activity @container) in order to safely make the fragment swap. The new container has a margin in order to always show the toolbar. Note that now, except SlideMenuFragment, all fragment should be attached to fragmentContainer,
and should call "container.removeAllView()" to ensure that the previous fragment is completely removed from the container.